Remove legacy codes
